working through modularizing
